Implemented an indexed buffer management system to simplify batch instancing and management of shared/common shader buffers (e.g. Transforms, Materials), and automate creation/management of index LUT buffers. Fixed multiple bugs around Buffer offsets. Optimized platform buffer buffering to combine neighboring writes. Optimized DX12 default heap buffering by using a single shared intermediate resource per frame. Lots of cleanup and bug fixes.

+++ Source/Renderer/CommandList_DX12.h
@@ -153,7 +153,7 @@ namespace dx12
 			size_t intermediateOffset);
 
 		void UpdateSubresources(
-			re::Buffer const*, uint32_t dstOffset, ID3D12Resource* srcResource, uint32_t srcOffset, uint32_t numBytes);
+			re::Buffer const*, uint32_t dstOffset, ID3D12Resource* srcResource, uint64_t srcOffset, uint64_t numBytes);
 
 		void CopyResource(ID3D12Resource* srcResource, ID3D12Resource* dstResource);
 		void CopyTexture(core::InvPtr<re::Texture> const& src, core::InvPtr<re::Texture> const& dst);
